backend/requirements.txt
plaid-python
python-dotenv==0.15.0
sqlalchemy
psycopg2
alembic
pydantic
sqlalchemy_utils
mongoengine
sshtunnel
fastapi
uvicorn
bcrypt
pyjwt
asyncpg
email-validator
pytest
httpx